If the command passed to a classes `do_help` function is a command from that class and not an action, call the `do_help` from the command class instead. For example, the command `help users` actually calls `users help`.
If the second argument of a line received on a classes `lineReceived` is a command from that class and the first argument is an action of that command, tweak the line received as it would have been issued for that command's class and not the original one. For example, the command `help users add` would actually call `users help add`.

These changes are just to make the help system a bit better.
Any writing to the terminal is now done by a deferred.